+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 3.0-alpha9
 - Changed counts to round down
 - Saving and loading templates is now enabled
+- You can now toggle bank inclusion for single tracker objectives
 
 ## 3.0-alpha8
 - Restructured database